Fully featured
This is what I consider the minimum viable version of Auteur. The result is a working client that allows Authors to create .atr files that can then be picked up by that same client from a root directory and compiled into an up to date credits.txt

Next I would like to have launch options, and see about multi-platform support so the workflow with Auteur can be completely automated, regardless of your tech stack. I also want better error handling and more communication to the user.
-E